Fixed bug with RPi3 prompt and switch to fkms
I was using the wrong variable name for the if block that handles setting up the RPi3.  I fixed that bug and also switched the behavior to use the fkms driver in the case of an RPi3 with official display.  Although the real KMS driver does work, it inconsistently detects and sets up the overlay for the touch screen.  Attempts to force the overlay are also flaking causing the app to fail on a portion of startups due to no /dev/dri directory.

The primary reason to prefer the KMS driver is that all current/future development is happening there and supposedly the KMS driver is already faster than the old code.  But reliability/stability trumps performance in this case.